app/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
chardet==4.0.0
Django==3.1.5
django-admin-interface==0.14.2
django-appconf==1.0.4
django-calendarium==1.3.4
django-colorfield==0.4.1
django-easy-select2==1.5.7
django-filer==2.0.2
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-libs==2.0.3
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-nested-admin==3.3.3
django-nested-inline==0.4.2
django-polymorphic==3.0.0
django-select2==7.5.0
django-smart-selects==1.5.9
django-tables2==2.3.4
djangorestframework==3.12.2
easy-thumbnails==2.7.1
fdfgen==0.16.1
future==0.18.2
html5lib==1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
pdfjinja==1.1.0
pdfminer.six==20160202
Pillow==8.1.0
PyPDF2==1.26.0
pypdftk==0.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-monkey-business==1.0.0
pytz==2020.5
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.2
webencodings==0.5.1
xhtml2pdf==0.2.5
gunicorn==20.0.4